refactor: onboarding post change user-info (#99)
