+++ src/app/api/user-info/route.ts
@@ -59,3 +59,26 @@ export async function DELETE(request: Request) {
     return NextResponse.json({ message: "error" }, { status: 500 });
   }
 }
+
+export async function POST(request: Request) {
+  try {
+    const supabase = createClient();
+    const postData = await request.json();
+
+    console.log(postData);
+
+    const { error } = await supabase
+      .from("userinfo")
+      .upsert({ ...postData, onboardingCheck: true });
+
+    console.log("errorororror", error);
+
+    if (error) {
+      return NextResponse.json({ message: error.message, status: 400 });
+    }
+
+    return NextResponse.json({ message: "ok", status: 200 });
+  } catch (err) {
+    return NextResponse.json({ message: "error", status: 500 });
+  }
+}
